dbt/plugins/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lakehouse-dbt-spark-patch"
version = "0.1.0"
description = "Patch dbt-spark Thrift/PyHive para Glue (database + Iceberg v2); carregado via sitecustomize."
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = ["sitecustomize"]
